Added validation on activate script to avoid running it from the wrong folder

+++ activate.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
+if [ ! -f activate.sh ]; then
+    echo "This script should be run from the .dotfiles dir"
+    exit 1
+fi
+
 cd ~
 ln -s .dotfiles/gitconfig .gitconfig  
 ln -s .dotfiles/githelpers .githelpers 
 ln -s .dotfiles/gitignore .gitignore  
-ln -s .dotfiles/gitmodules .gitmodules 
 ln -s .dotfiles/vim .vim        
 ln -s .dotfiles/vimrc .vimrc        
 ln -s .dotfiles/irbrc .irbrc
